lib: hal: fspi: Increase the maximum configurable number of DLL interfaces

+++ lib/hal/inc/hal_fspi.h
@@ -129,7 +129,7 @@ HAL_Status HAL_FSPI_MaskDMAInterrupt(struct HAL_FSPI_HOST *host);
 HAL_Status HAL_FSPI_UnmaskDMAInterrupt(struct HAL_FSPI_HOST *host);
 HAL_Status HAL_FSPI_XmmcSetting(struct HAL_FSPI_HOST *host, struct HAL_SPI_MEM_OP *op);
 HAL_Status HAL_FSPI_XmmcRequest(struct HAL_FSPI_HOST *host, uint8_t on);
-HAL_Status HAL_FSPI_SetDelayLines(struct HAL_FSPI_HOST *host, uint8_t cells);
+HAL_Status HAL_FSPI_SetDelayLines(struct HAL_FSPI_HOST *host, uint16_t cells);
 HAL_Status HAL_FSPI_DLLDisable(struct HAL_FSPI_HOST *host);
 uint32_t HAL_FSPI_GetXMMCStatus(struct HAL_FSPI_HOST *host);
 uint32_t HAL_FSPI_GetMaxIoSize(struct HAL_FSPI_HOST *host);
